Adding calture info

+++ Webpay.Integration.CSharp/Webpay.Integration.CSharp/Order/Row/Credit/NewCreditOrderRowBuilder.cs
@@ -6,7 +6,7 @@
         public long UnitPrice { get; set; }
         public decimal Quantity { get; set; }
         public decimal VatPercent { get; set; }
-        public decimal DiscountAmount { get; set; }
+        public long DiscountAmount { get; set; }
         public decimal DiscountPercent { get; set; }
         public string Unit { get; set; }
         public string ArticleNumber { get; set; }
